Move helper functions to utils folder
